refactor(drug-discovery): dedupe _make_payload / DTO helpers via execution.py
uniprot_discovery.py's _make_payload, _optional_float, and
_execution_outputs_dict were byte-identical to structure_report.py
(and konnektor.py for _make_payload), which SonarCloud flagged as
new-code duplication over the 3% gate. Hoist the shared bodies into
execution.py as _default_execution_payload / _execution_outputs_dict
/ _optional_float and have all three call sites use them; behavior
is unchanged (pure extraction), verified by the full local suite.
